fix Arbitrary AssociatedFile to not crash when LANG=C
Even letting through things that Data.Char.generalCategory said
wereUppercaseLetter caused the crash. Apparently what's going on is
that, in LANG=C, it does not expect to find unicode chars in a String,
except presumably ones that are surrogates.

But ascii is good enough to test the things we need to test about
associated files.