updates for new cabal
The old make fast machinery doesn't work at all with it, but it seems
actually reasonably fast now to just disable optimisation, so made make
dev do that. And it sticks once used, which is handy.

Removed hdevtools hack, which won't work. There are ide integrations
that really solve the problem that hack was working around.

Misc fixes for dist-newstyle location.